Fix compilation errors after merging main into hedging branch
Reconcile hedging code with the new crate::error::CosmosError surface introduced on main:

- operation_pipeline: unify hedging helpers (finalize_hedge_attempt, perform_single_attempt, application_cancelled_error, transient_outcome_error, try_advance_after_both_transient, HedgedRaceResult variants, futures bounds) and all azure_core::Result<TransportResult> sites to crate::error::Result; switch enforce_deadline_or_timeout calls to value-pass-and-return idiom; destructure cosmos_headers (was headers); replace build_http_error import with build_service_error; rewrite test helpers and assertions (http_result u16 cast + 3-arg from_http_response, harvest_remaining_attempt fakes, drop err.kind() asserts in favor of err.status().status_code()).- retry_evaluation: add std::sync::atomic::Ordering import and make build_service_error pub(crate).- hedging_eligibility: cast sub-status to u16 for CosmosStatus::with_sub_status.- tests/in_memory_emulator_tests/hedging: switch helper return type to crate::error::Result and drop unused HttpClient import.

+++ sdk/cosmos/azure_data_cosmos_driver/src/driver/pipeline/hedging_eligibility.rs
@@ -317,7 +317,7 @@ mod tests {
     fn status(code: u16, sub: Option<u32>) -> CosmosStatus {
         let mut s = CosmosStatus::new(StatusCode::from(code));
         if let Some(v) = sub {
-            s = s.with_sub_status(v);
+            s = s.with_sub_status(v as u16);
         }
         s
     }
